fix: test bugs due to typing

+++ src/ss/system/_system_dynamic.py
@@ -20,9 +20,7 @@ class ContinuousTimeSystem(System):
             super().__init__()
             if noise_covariance is None:
                 noise_covariance = np.zeros((dimension, dimension))
-            self._noise_covariance = np.array(
-                noise_covariance, dtype=np.float64
-            )
+            self._noise_covariance = np.array(noise_covariance)
             self._dimension = dimension
             self._name = name if name is not None else "noise_covariance"
             self._validate_functions.append(self._validate_shape)
